Add accessibility testing via aXe
These tests will look for '[data-module="test-a11y"]' and run axe against them,
it'll then throw an error to be caught in dev tools or better in an applications'
integration tests.

Adds Jasmine as a development dependency to test this behaviour, we add require jasmine to the dummy application
so that the default rake build will execute successfully.