Move guide CSS into parent file
Static behaves strangely when autoloading govuk_frontend_toolkit and
importing the component_guide.

Somehow the component guide sass file was loaded outside of the scope
of the toolkit, so did not have access to things like $gutter and
core-19.

By moving into a single file we avoid this problem. Although the
initial cause is still unclear.