Disable duplicate ID rule in axe tests
Components sometimes contain IDs. These get repeated on the component
guide page. This repetition is not representative of actual usage,
leading to a false positive accessibility failure.

This specifically affects the document_footer component in static where
we rely on a specific component ID to link to from elsewhere in the
page – so it can’t be dynamic and unique with each use.