Add uglifier as a development dependency
This allows us to run asset precompilation in rails production mode
which is a slightly better test of asset precompilation than running it
in dev or test mode which has differences to production (digests and
compression mostly).

We add the version that you would get with rails 5.0.0.1 because that's
the version we've got as a dependency.