Support components accepting blocks
ERB's render method accepts a block, which is useful for components which
embed multiple components for a specific purpose (eg. a content list with
a back-to-top link).
The component guide should be able to render the docs for this type of component
so this commit adds the ability to configure the component docs with the
examples[data][block] parameter, this will render an example block in
the guide and render how to call the component with a block.