Remove doubly included stylesheets
The component CSS is already included into the component guide, because
the component guide pulls in the main CSS of the application (which
includes the components if configured correctly).

To make sure that new components get added to the dummy app, 2 import
files are now provided that include all the styles. Applications can
also use this if they don't want to import all the components
individually.

This isn't in the README yet because I'm not sure where to place it.
We'll look at restructuring the installation instructions for the gem.