Update related_navigation element ids
We have come across an issue in the component guide where the page for
`related_navigation` won't render due to duplicate id errors. After
investigation we've discovered that for some reason this component is being
rendered twice by the component guide, which is causing this error. There is
currently a discussion around the efficacy of the headers in this particular
component from an accessibility point of view, so until a conclusion is reached
from these discussions, we implement a temporary fix where a random integer is
appended to the ids.