Remove "report a problem" header
We can't use this because otherwise apps that are upgrading to the
latest version of slimmer
(https://github.com/alphagov/slimmer/pull/215) stop working because the
`report_a_problem ` header isn't allowed anymore. It would be better to
bump the slimmer version here too, but we can't do that yet because it
would require apps to upgrade before adding the new feedback component.

This will cause the "Is there anything wrong with this page" to show on
the component guide. It will disappear once we upgrade slimmer.

https://trello.com/c/eKtXGtBD