Remove Related Navigation Sidebar Helper dependencies
When moving Related Navigation Sidebar Helper from `govuk_navigation_helpers`
to this gem, we found that it had a couple of dependencies that lived in
`govuk_navigation_helpers`. We know that `govuk_navigation_helpers` is being
deprecated, so we don't want to introduce it as a dependency in this gem.
Instead we incorporate the methods that Related Navigation Sidebar Helper
requires from its dependencies into this helper.