Make input work on older layouts
The bank holidays page is using an older layout, which includes the
button CSS. This contains a style rule on the `input` element, rather
than a class.

https://github.com/alphagov/static/blob/79d29b6a4221874ead27e67d007704a6
be166a57/app/assets/stylesheets/helpers/_buttons.scss#L91-L122

To make the styles work on these layouts we need to make the component
more specific (unfortunately). We also reset the margin to override the
styles from buttons.css

https://trello.com/c/8TC3AEms