Update document list with description
We want to make use of the document list component on topic pages. For mainstream services, we want to show a description of the document rather than metadata.

In the future, we may also want to supply additional metadata for each document, such as organisation and collection the document belongs to. This PR refactors the component to stop it being limited to only updated date and document type