Fix mainstream browse page link bug
We found a bug whereby grandparent mainstream browse pages weren't being added
to the "topics" part of the `related_navigation` payload. This updates a couple
of methods where the content item's keys weren't being handled correctly and
adds a test to check for the presence of parent and grandparent mainstream
browse pages.