Add option to removed side padding if full width
In this commit I have added the option to pass a flag to the component
to indicate if the header component will cover the full width of the
page. If so, the left-right padding is removed for the header so it’s 
content can line up with the rest of the content on the page.

I also added a bottom margin of 30px to the component.