Make changes released in 5.5.3 backwards compatible
Because I didn't put in a changelog, 5.5.3 got accidentally released as
a minor version. This makes the changes backwards compatible so that
apps don't have to update.

We'll do the updating at a later time, when there are more breaking
changes to ship at the same time. That hopefully minimises the impact.