Add contextual navigation preview
This is a page in the dummy app which is available on Heroku. It will
allow you do see what the contextual navigation for any page on GOV.UK
would be if it were using the components. There's a hard coded list of
examples which we can refer to. I hope that that gives us more
visibility on the different types of navigation that exist.