Tweak spacing for using inverse header on html pub
As the publication header and inverse header components are so similar, it makes sense to consolidate the two into one component.
To do this, we need to make a few spacing changes to the inverse header component so it works within the HTML publication layout.
This change should not break any current uses of the inverse header, e.g: on topic pages