Show "normal" navigation in addition to step by step
Currently, when the page is tagged to 2,3 or 4 step by step lists, the
sidebar will only show links to those steps by steps and not the links
to navigation that would otherwise be there.

This is a theoretical change because there aren't any pages tagged to
more than 1 step by step currently.