Simplify RelatedNavigationHelper
This changes the output of the RelatedNavigationHelper from an array to
a hash. If I remember the conversation with Steve correctly, the array
was chosen to because the ordering is significant. In practice, Ruby's
hashes are pretty much ordered. This means we can do away with some of
the extra loops and make the data structure simpler.