Simplify logic around related links in the sidebar
Currently, our logic whether to show the taxonomy sidebar or the new
related navigation sidebar is to check if the page is tagged to
mainstream browse. This _almost_ overlaps with the existence of related
links, but not entirely. This meant we introduced the taxonomy sidebar
for the travel advice pages too, which we fixed by adding another
conditional in
https://github.com/alphagov/govuk_publishing_components/pull/264.

This PR switches the logic to what it probably always was supposed to
be: we show related links when 1) there are related links OR 2) the page
is tagged to mainstream browse.

Because the resulting behaviour is the same, no tests should have to
change.

https://trello.com/c/rWqoHcOu