Fix spacing issue in radio buttons
This component doesn't work in some situations. The clickable area of
the radio button is about half the size that it should be.

According to @andysellick:

> Seems to be that the component is inadvertently relying upon some
core styles for inputs, which includes font-size: inherit. As a test,
if you replace these two lines with that, the input comes out at the
right size. Suggest we make an adjustment to the component to define an
explicit font size so this works in isolation - might be able to move
the font size declaration for the label higher up so it applies to the
whole component.