Revert "Merge pull request #302 from alphagov/add-canonical-url-meta-tag"
This reverts commit b53fa52329520c848ac0f9577f9f1fc500d9d7ed, reversing
changes made to 8f9ed427d7fd6ae2fab198825b1055b77291fd7a.

After adding this we realised that it would be better placed for the
individual frontend applications to add the canonical tag as they work
closer to the content and would understand how to generate different
canonical URLs depending on document type and various other factors.

Although we could choose to leave it in here, we felt it would be
better for apps to be explicit about the tag rather than having a
default which might cause more harm than good.