Stop inserting &nbsp; between last 2 words
This feature was added when the component lived in Static. It replaces
the space between the last 2 words with a non-breaking space, so we'll
never have a dangling word in a paragraph.

This PR removes the feature because:

1. This component was recently moved into the gem. While converting
applications to use it, I find that it turns out to be very hard to
maintain this feature without jumping through a lot of hoops. For
example, a [simple test like this][1] doesn't work, because
`@description` doesn't have the nbsp, and the actual page does. This is
*very* surprising, even to me -- and I ported this feature so I
should've known.

2. The Sanitisation code is pretty inscrutable, and introduces
potential security vulnerabilities.

[1]:
https://github.com/alphagov/government-frontend/blob/23ab4380b09e1e76bfb
ab70079530900a1f3e7dc/test/integration/specialist_document_test.rb#L34