Use h2 as the default for `heading` component
The heading component currently is a h1 by default.

However, most common usage of this component is to add subheadings in a
page, not a page title (for which we tend to use the `title` component).

This changes the default to a `h2` because it's more intuitive and
apparently more expected - it fixes a number of duplicated h1s in
government-frontend and collections.

https://github.com/alphagov/government-frontend/blob/69c11af5b67c1b27bce
047372fb2e899fee3b873/app/views/content_items/_publication_inline_body.h
tml.erb#L12-L15

https://github.com/alphagov/government-frontend/blob/69c11af5b67c1b27bce
047372fb2e899fee3b873/app/views/content_items/consultation.html.erb#L40