Tweak context spacing on document list component
The spacing on the document list component was originally set on the context element itself.
This meant that on smaller screens, the context text appeared to be indented from the left when it wrapped onto a new line.
By moving the margin to the link title, the context text will appear in line with the link title (on the left) if it wraps.