Update document list documentation
The document list component was going to be used on the new /government/organisations pages in Collections.
However, we've decided that using this component will introduce more limitations.
We therefore need to update this documentation in case it misleads people in the future.

I'm going to leave the component functionality in place though as it still may be useful to org pages/other teams.