Add GOV.UK Frontend
This commit adds GOV.UK Frontend
(https://design-system.service.gov.uk/) to the gem.

When we run `npm install` npm will install the library into
`node_modules`. This is ignored in git to avoid us having to commit the
entire 15.000 lines of code in govuk_frontend. To include it in the
distributed version of the gem, the node_modules directory is added to
`s.files` in the gemspec. To make the files available during build, we
run `npm install` before the tests (see Jenkinsfile).

PRs that actually make use of GOV.UK Frontend will come separately.

https://trello.com/c/R7Ija8dR/16-create-layout-system-for-admin-applicat
ions