Add component for the basic page HTML
This component outputs just the basic HTML for a page. It has `<head>`
and `<body>` tags, title tag and includes its own CSS. So preview the
whole thing, there's also a special page on /admin in the component
guide.

In future PRs we'll add a header, footer and the rest of the page
furniture.

https://trello.com/c/R7Ija8dR