Add footer component
This implements the footer component from GOV.UK Frontend.

Things of note:

- Even though we don't have custom CSS, we still create a CSS file for
this component. This is mainly so it's easy to discover for a developer
where the styles come from.
- When the component is `part_of_admin_layout`, we'll show a warning
banner. This hopefully makes it clear that the component can't be used
on the public frontend (because GOV.UK Frontend isn't available).

https://trello.com/c/k5agy6jw