Add test that components have the correct class
This adds a test to make sure that every component has  a specific
formatted classname. This will encourage consistency and codifies that
we'll also add the `gem-c-` class to GOV.UK Frontend-powered
components. This makes sure that the browser extension can highlight
and link all of our components.

There are a couple of components that have weird class names, so I've
added those to `skip` (meaning they should be fixed). For example, the
class name for the previous_and_next_navigation component is
`gem-c-pagination`, not `gem-c-previous-and-next-navigation` as one
might expect.

There are 3 components that don't need a class name, they're in
`not_applicable` (which means they will not be fixed).