Fix the step by step title sizing and spacing on mobile when it appears as a footer
## Changes
- changes margin bottom on step by step nav related from $gutter * 1.5 to $gutter-half for both desktop and mobile
- adds font size bold-24 for mobile views

## What
On mobile when the step by step nav appears as a footer the sizing of the step by step title is too small and there's a lot of whitespace between the title and the steps beneath.

We should bump up the size of the title and close up the space a little. Check-in with Kate if you need guidance on sizing and spacing.

## Why
One of our goals this quarter is to optimise the user-facing step by step pattern.

Ticket: https://trello.com/c/zlQ0Go3P/700-fix-the-step-by-step-title-sizing-and-spacing-on-mobile-when-it-appears-as-a-footer