Enable bold text by default in govspeak
Certain types of guidance have tested positively with inline
bold text (eg. https://www.gov.uk/guidance/connect-to-govwifi-using-an-android-device)
Instead of enabling bold text as by exception we'd like to see
the effect of enabling entirely.
If nothing else it's good to talk through why this is a good or bad
idea and what the effect of permitting bold text has and will be on
content managed text.