Relevant step by step should now be open inside simple smart answers
This changes the assignment of requested path of document type
"simple_smart_answer" to use the base path of the content item rather
than the requested based path.

This was done so that the active link would show correctly on
simple smart answers.

The reason why this cannot be done for all pages is because
some content items have their base path different to the
actual path - ie. guidance pages.

Ticket: https://trello.com/c/SaxKuvqK/726-relevant-step-by-step-should-be-open-inside-simple-smart-answers-showing-the-dot-on-the-task-m