Only show the part of title for the active step by step
If we don't add this check then the titles all the step by steps
the content item is part of will show above the expanded step by
step and this will confuse people.

We will show the titles for the other step by steps in an "also part
of" section that will add in a later commit.