Avoid using an optional block in success alert component
We currently have a really good bug blocking
https://github.com/alphagov/content-publisher/pull/220 from being
merged.

The success alert component intends to check if there's a block given
to the alert component. It does so by checking `yield`. This works in
tests, but in the real world there will *always* be a return value of
`yield` - either the given block, or the *entire layout of the
application*.

Unfortunately, we have no way of determining which `yield` we're
talking to, so I think we won't be able to have optional blocks.

https://trello.com/c/i2Fi5qNQ