Refactor header component to provide options for head or list
When you're on a page that is part of two step by steps and one is expanded the other is displayed underneath as a heading but has no associated content.

When this happens the current orphaned header should be render as a list.

We should provide the option to display this as a list.

Ticket: https://trello.com/c/ul9DX8yu/856-refactor-header-component-to-provide-options-for-head-or-list