Move script elements to just before closing body tag
This moves the script elements from the HTML head to be just before the
closing body tag.

The reasons for this change are:

- This is a standard practice for not hindering a web page rendering
  with blocking scripts
- This allows to embrace the script initialisation approach used by
  [GOV.UK Frontend][1]

The need for this is triggered by Content Publisher using the GOV.UK
Frontend initialisation approach and thus the components do not work.

As far as I can tell this shouldn't cause any problems for existing apps
as all of their JS is initialised within `$(document).ready` functions
so will still execute at the same time.

[1]: https://github.com/alphagov/govuk-frontend/blob/master/docs/installation/installing-from-dist.md#2-include-resources