Make data from related_to_step_nav property to be accessisible
This is required because when a content item has step by step disabled
it does not have the step_navs property but has the
step by steps within related_to_step_navs.

This is needed for the personalisation work that will be done to show step
by steps on pages where we know the user is actively interacting with a
specific step by step.