Drop the admin_styles IE8 stylesheet
The admin_styles_ie8 stylesheet contains all of the styles of GOV.UK
Frontend for IE8. This creates a complicated situation as we now include
all of GOV.UK frontend in the all components stylesheets and pull bits
of GOV.UK Frontend without regard for IE8.

After discussion with @alex-ju we decided the most realistic course of
action for GOV.UK is to drop these particular styles and only consider
enabling this if we find we have an explicit need to fully support IE8
on a project.