Update govspeak lists for no margins
This changes these so they are identically styled whether the static
reset stylesheet is present or not.

I did start these by going down the route of embracing govuk-frontend
with an approach of:

```
  ul {
    @extend %govuk-list;
    @extend %govuk-list--bullet;
  }

  ol {
    @extend %govuk-list;
    @extend %govuk-list--number;
  }
```

And this did work with a few modifications, but it felt like too much of
a far reaching change to apply at this point.