Style h1 inside search component label
This applies the static reset to a h1 embedded into the search component
label.

It's not ideal staring a bare element here, we really should just avoid
passing in a h1 without a class but this is something we already have
going on, so doing this will maintain backward compatibility and have
the component not look broken with static.