Add javascripts only if we don't have slimmer
We got so far without any need to reference slimmer at all, but
unfortunately including jQuery a 2nd time breaks the magna-charta jQuery
plugin.

The easiest way I could think to get out of this was to do a check for
slimmer and include an extra javascript file if it doesn't exist.