Don't use Slimmer even if apps have Slimmer
This makes the component guide run consistently without Slimmer on all
apps that use it, rather than some having Slimmer and some not having
it.

The intention of this is so that the experience of adding a component to
your app and to the overall guide is the same and that they can't rely
on an underlying static introduced dependency.