Expose the high resolution image in meta tags
As described in
https://github.com/alphagov/govuk-content-schemas/pull/827, we
currently expose the image from the page to social media scrapers using
this component. This image is often quite small, and unsuitable for
services like Twitter.

Whitehall will soon be sending a higher resolution image in
`details.image`. This PR makes sure that we use it if it's present.

https://trello.com/c/eqcMykln