Handle smaller images in image card large
This tweaks the CSS for the image card large example so that it now handles smaller images more gracefully. Previously, there would have been a large gap between the small image and the text. Now, the text expands to fill the rest of the space not taken up by the image.