Add accessible autocomplete componenet
This commit adds a new accessible autocomplete component to the publishing components, using the existing accessible autocomplete at https://alphagov.github.io/accessible-autocomplete.

The new component here adds a label alongside the autocomplete, in addition to support for tracking if this is specifically enabled when creating the component.

No JS tests are included as they would be testing the functionality of the existing autocomplete component (which can be seen at https://github.com/alphagov/accessible-autocomplete/tree/master/test).

Co-authored-by: Mahmud Hussain <mahmud.hussain@digital.cabinet-office.gov.uk>