Make h1 in legend optional
The default legend shouldn't have a h1 in it so
the component can be used in pages where it isn't
the only form control.

This brings it into line with the checkboxes
component from the GOV.UK Design System version:

https://design-system.service.gov.uk/get-started/labels-legends-headings/

I took the name of the flag from there and changed
it to fit our naming conventions.