Allow textarea to take a block for extra content
This allows embedding extra elements into the form group of this
component. I'm not too sure this is an ideal path since this could grow
to needing elements before it as well, but as it's within a form-group
there's going to be a need for more elements inside it.

Alternative approach would be to abstract to a textarea and form-group
components.