Allow embed option in component docs
The embed option allows for component examples to be illustrated within
contextual HTML. It is intended for components that do not have a visual
presence without some other HTML. The Content Publisher application has
a number of these, such as Contextual Guidance and Input Length
Suggestor, and is likely to have another for modals. These cannot be
viewed inside the component-guide and just appear as empty boxes.

This resolves this by allowing the ability to include an option of
`embed` in a component document file which contains the HTML that is
needed in the vicinity of the component eg:

```
name: Component
embed: |
  <button data-trigger-component=true>Click me</button>
  <%= component %>
examples:
  default:
```

This can also be applied on individual examples eg:

```
name: Component
examples:
  default:
  inside_div:
    description: This component behaves differently inside a div
    embed: |
      <div><%= component %></div>
```