Don't send the `user_agent` in survey form
We're currently adding the `user_agent` when the page is first
requested. This is wrong because the page will be cached, so that any
subsequent page requests will return the same HTML with the same form
fields - and the previous request's user agent.

This commit removes the user agent, because we don't actually seem to
be using it in the feedback application where it's sent to: the
EmailSurveySignupController creates a `EmailSurveySignup` object, which
sends the actual email. There's no usage of this `user_agent` anywhere.