Don't send user_agent in problem reports
This removes the `user_agent` from the problem report form that the
user sees when clicking "Is there anything wrong with this page".

Like https://github.com/alphagov/govuk_publishing_components/pull/698,
this is wrong because the page will be cached, so that any subsequent
page requests will return the same HTML with the same form fields - and
the previous request's user agent.

We can remove this field, because it turns out that the feedback
application merges the sent `params` with `technical_attributes`, which
includes the correct user_agent.

https://github.com/alphagov/feedback/blob/6f74a9a2887ceff3302a4e8f425aff
715dbc3e4a/app/controllers/contact/govuk/problem_reports_controller.rb#L
9