Stop showing taxonomy breadcrumbs when viewing specialist documents
This commit stops the contextual breadcrumbs component from showing taxonomy breadcrumbs when the content being viewed is a specialist document. This allows the content item's finder to be displayed instead, for which the logic already exists in the `breadcrumbs` method of `contextual_navigation.rb`. By making this change, we allow users to more easily navigate to the correct specialist finder if they come from somewhere other than a specialist finder.

Note that if `prioritise_taxon_breadcrumbs` is specified and set to `true` as part of instantiating the contextual breadcrumbs component, it is possible to override the above behaviour.

Solo: @karlbaker02