Support rendering raw HTML in notice component
Previously only paragraph text or Govspeak could be passed to this
component. This adds a 'description_text' option as an alternative,
which is not nested in any other tag.

As with the Govspeak component, it seems better to require the consumer
of this component to specify that the HTML they're passing-in is safe,
so it was necessary to specify this in the tests.