Update pull request template
There are 32 PRs with "THIS PR NUMBER" in the description, and 38 with "Remember to update the CHANGELOG if your change needs to be listed in the next version of the gem". Something isn't working with the PR templates.

This removes the link to Heroku, because that'll be added automatically by Heroku. Also hides "Remember to update the CHANGELOG" in a comment tag so only the PR author sees it.