Rename notice description_html to description
Previously the notice component would accept a block in order to set the
description_govspeak variable. Turns out this isn't currently used, and
rendering Govspeak within a component is not consistent with the others.

This changes the notice component to accept a block that will now set
the description_html parameter. As this change increases the importance
of this parameter, it makes sense to rename it to just 'description'.