Add JS dependencies for Youtube videos in Govspeak
This adds in two files that were in different GOV.UK repos:
jquery.player.min.js from govuk_frontend_toolkit and
jquery-ui-1.10.2.custom.js from static.

Both of these are now rather old dependencies by modern JS standards so
it's a bit of a shame to keep them around. But until we have a new way
to embed videos this is the only route.