Convert Youtube links to video embeds in Govspeak
This ports code from static [1] which allows the embedding of Youtube
videos into Govspeak. Once this work is complete it is intended that the code
in static can be removed.

It sets a Govspeak embed to have a data-module attribute that will cause
it to be initialised when the page is ready (thanks to a module approach
used from govuk_frontend_toolkit), this will then check if any links
need converting to Youtube videos.

The function for extracting the video id from a URL is a class method
rather an instance method of the class. This is so that this method can
is available for apps that are allowing the embedding of videos, they
can use this to check whether a URL will work or not and provide user
feedback.

Static previously converted links into Youtube videos by calling a
function on all elements with a govuk-govspeak class on them. Moving
this to govuk_publishing_components will mean this functionality is lost. To
support this there is a `GOVUK.GovspeakYoutubeLinkEnhancement.initLegacy` function
that static can call to maintain existing behaviour without needing a
duplication of this file.

There is a risk that both the static initialisation and this component
initialisation might both occur. I've put a selector in to try mitigate
this and hopefully, given links are removed, nothing too bad can happen
on a subsequent initialisation.

[1]: https://github.com/alphagov/static/blob/233964da3622de0b4d9ee9c5e52b6849ab510513/app/assets/javascripts/govuk-component/govspeak-enhance-youtube-video-links.js