Include magna-charta vendor lib
This copies this file over from static [1] with the intention that once
this work is completed we can remove it from static.

Magna-charta is a GDS authored jQuery plugin that is used to enhance
HTML tables into barcharts. See https://github.com/alphagov/magna-charta
for more info. It's a rather old library, at the time of writing it's
almost 4 years since it's last change.

[1]: https://github.com/alphagov/static/blob/233964da3622de0b4d9ee9c5e52b6849ab510513/app/assets/javascripts/libs/govuk-component/govspeak-magna-charta.min.js