Port Govspeak barchart enhancement
This ports the Govspeak barchart enchancement from static [1] to
govuk_publishing_components. With the intention that once the work is
completed in porting it, the dependencies of this can be removed from
static.

Similar to the GovspeakYoutubeLinkEnhancement, this also provides an
initLegacy function that is intended for apps like static to call so
they can enhance govspeak that is embedded without using the govspeak
component.

This embed is quite vulnerable to being run multiple times - result
being it tries to create a barchart inside the existing one. As on it's
initial initialisation it creates a new element with the same selector
(.js-barchart-table) as what we use to initialise it. Because of this
I've added two `:not` selectors to the initialisation to hopefully avoid
this running on items twice. As part of this protection a class of
`js-barchart-table-init` is added so we know it has been run.