Rather reckless fix bug in minified JS
There is a bug here that can be seen in easier in the unminified code on
github [1]. Which looks like so:

```
this.player = YT.Player('player-' + player.config.id, getOptions(player));
```

This has a variable, player, and a function, getOptions, that are both
not defined. These instead should be:

```
this.player = YT.Player('player-' + this.config.id, this.getYTOptions());
```

which is the change this commit does.

The project that provides this seems long abandoned with no updates in
over 4 years so seems of little value to resolve it there.

[1]: https://github.com/nomensa/Accessible-Media-Player/blob/8f4bc98ad6b45278890987959dde5c1bb83b1217/core/javascript/youtube_player.js#L83