Test embedded Youtube links with mocks
Testing Youtube links seems to be a dangerous, delicate game where
underlying JS that is brought in can raise errors. These only seem to
occur when we embed HTML into the actual page rather than just creating
an element in JavaScript.

The only time testing that relies on HTML being on the page is for
initLegacy, thus we can test this with a mock so the Youtube is never
actually embedded.

If it transpires the other tests are flaky we may need more mocks or to
drop the tests altogether.