Remove legacy init methods for govspeak JS enhancements
After having a dig around all the GOV.UK frontend apps I've failed to
find any places where we're rendering govspeak without
govuk_publishing_components - the only one I can find is potentially
smart-answers, but that doesn't seem to use the govuk-govspeak css class
nor do the files have barcharts or Youtube videos.

Rather than rebasing to remove this code I've decided it's better kept
in the log in case we do need to pull it back for a case I've not found.

So I've decided this can be deployed without the need for these
initLegacy methods. It'll be a challenging deploy as we'd want to deploy
a bunch of apps with updated govuk_publishing_components at a very
similar time to deploying static to remove the enhancement, but there
doesn't seem to be a particularly easy way around this.