Add workaround to provide vertical spacing for input items on narrow containers
Ideally we would consistently add margin-bottom to these input items, but that then when we have an error state it will result in the left border exceeding the height of the content. Added top margin in a similar approach with how we deal with hints and error messages on this component.