refactor failing a11y specs
Two a11y specs started failing as a result of switching from poltergeist
to headless chrome. The reason for this is that the expectation was for
poltergeist to raise a JS errror. Unfortunately Chrome does not have a
direct replacement for detecting js failures but it does expose any
browser console.errors/logs which is what this refactor uses to ensure
that there are was js errors.