Suppress subscription components heading
For some unknown reason this component includes a h2 heading by default.
This means that if a service uses it unexpectly it will include the heading.
It also stops a team from using this component anywhere else on a page.

By using the 'hide_heading' option, it will suppress the heading when
the component is being rendered.